requirements.txt
matplotlib==3.6.2
numpy==1.23.5
Pillow==9.3.0
tqdm==4.64.1
wandb==0.13.5
torch>=1.7.0
torchvision>=0.8.0
scikit-learn>=0.24.0
